Basic timebase via LAPIC timers; clang toolchain
- Created basic drivers for the local APIC timers; we have to guess the
  frequency of the timer clock by timing it against the PIT. I have no
  idea how reliable this is on real hardware but it's the only way we
  can get it until later CPUID extensions :(

- Added division intrinsics provided by x86 arch code, rather than
  relying on compiler runtime libraries

- Fixed a few incorrect format args (gcc really is fucking dumb)

- Reworked clang toolchain files to actually work; you'll have to build
  a toolchain yourself and adjust the paths, but it does work.

- Reimplemented queues as an adapter over linked list

- Added linked list template class

  Currently, there's some sort of issue in the context switching
  routines that causes them to not work quite right when compiled with
  clang. This needs to be figured out.

+++ kernel/arch/x86/include/arch/atomic.h
@@ -18,14 +18,14 @@
 
 /// Atomic 32-bit exchange
 static inline __attribute__((always_inline)) uint32_t xchg_32(volatile void *ptr, uint32_t x) {
-    asm volatile("xchgl %0, %1" :"=r" ((uint32_t) x) :"m" (*(volatile uint32_t *) ptr), "0" (x)
+    asm volatile("xchgl %0, %1" :"=r" (x) :"m" (*(volatile uint32_t *) ptr), "0" (x)
                  :"memory");
     return x;
 }
 
 /// Atomic 16-bit exchange
 static inline __attribute__((always_inline)) uint16_t xchg_16(void *ptr, uint16_t x) {
-    asm volatile("xchgw %0,%1" : "=r" ((uint16_t) x) :"m" (*(volatile uint16_t *) ptr),
+    asm volatile("xchgw %0,%1" : "=r" (x) :"m" (*(volatile uint16_t *) ptr),
                  "0" (x) : "memory");
 
     return x;
